+++ CHANGELOG.rst
@@ -75,6 +75,10 @@ General
   hosts. The :ref:`debops.apache` Ansible role will use this module instead of
   the original one.
 
+- The :command:`debops exec` command can be used to execute Ansible modules
+  against hosts in the project directory; this is a wrapper for the
+  :command:`ansible` command.
+
 :ref:`debops.avahi` role
 ''''''''''''''''''''''''